+++ applications/fishing-map/src/features/map/popups/ContextLayers.tsx
@@ -52,7 +52,7 @@ function FeatureRow({
   const gfwUser = useSelector(isGFWUser)
 
   const handleReportClick = useCallback(
-    (ev: React.MouseEvent<Element, MouseEvent>) => {
+    (ev: React.MouseEvent) => {
       context.eventManager.once('click', (e: any) => e.stopPropagation(), ev.target)
       if (onReportClick) {
         onReportClick(feature)
@@ -240,7 +240,7 @@ function FeatureRow({
       </div>
     )
   }
-  return <div key={`${feature.value || gfw_id}`}>{feature.value}</div>
+  return <div key={`${feature.value || gfw_id}`}>{feature.value ?? feature.title}</div>
 }
 
 type ContextTooltipRowProps = {