+++ applications/vessel-history/src/features/map/controls/MapControls.tsx
@@ -11,6 +11,8 @@ import { selectDataviewInstancesByType } from 'features/dataviews/dataviews.sele
 import LayerSwitch from 'features/workspace/common/LayerSwitch'
 import { selectFilterUpdated } from 'features/vessels/activity/vessels-activity.selectors'
 import EventFilters from 'features/event-filters/EventFilters'
+import DataAndTerminology from 'features/data-and-terminology/DataAndTerminology'
+import ActivityDataAndTerminology from 'features/profile/components/activity/ActivityDataAndTerminology'
 import styles from './MapControls.module.css'
 
 const MapControls = ({
@@ -38,7 +40,7 @@ const MapControls = ({
     uaEvent({
       category: 'Vessel Detail ACTIVITY or MAP Tab',
       action: 'Open filters',
-      label: JSON.stringify({tab: 'MAP'})
+      label: JSON.stringify({ tab: 'MAP' }),
     })
     setIsOpen(isOpen)
   }, [])
@@ -53,11 +55,22 @@ const MapControls = ({
 
   return (
     <Fragment>
-      <EventFilters tab="MAP" isModalOpen={isModalOpen} onCloseModal={(isOpen) => setModalOpen(isOpen)}></EventFilters>
+      <EventFilters
+        tab="MAP"
+        isModalOpen={isModalOpen}
+        onCloseModal={(isOpen) => setModalOpen(isOpen)}
+      ></EventFilters>
       <div className={styles.mapControls} onMouseEnter={onMouseEnter}>
         <div className={cx('print-hidden', styles.controlsNested)}>
           {extendedControls && (
             <Fragment>
+              <IconButton
+                type="map-tool"
+                icon={filtered ? 'filter-on' : 'filter-off'}
+                size="medium"
+                tooltip={t('map.filters', 'Filter events')}
+                onClick={() => setModalOpen(true)}
+              />
               <IconButton
                 icon="layers"
                 type="map-tool"
@@ -72,13 +85,14 @@ const MapControls = ({
                   })
                 }}
               />
-              <IconButton
-                type="map-tool"
-                icon={filtered ? 'filter-on' : 'filter-off'}
+              <DataAndTerminology
+                containerClassName={styles.dataAndTerminologyContainer}
                 size="medium"
-                tooltip={t('map.filters', 'Filter events')}
-                onClick={() => setModalOpen(true)}
-              />
+                type="map-tool"
+                title={t('common.dataAndTerminology', 'Data and Terminology')}
+              >
+                <ActivityDataAndTerminology />
+              </DataAndTerminology>
               {showLayersPopup && (
                 <Modal
                   isOpen={showLayersPopup}
@@ -104,22 +118,20 @@ const MapControls = ({
                             />
                             {layer.id === 'context-layer-mpa' && (
                               <div>
-
                                 <IconButton
-                                  icon='info'
-                                  type='default'
+                                  icon="info"
+                                  type="default"
                                   size="small"
                                   className={styles.infoIcon}
                                   tooltipPlacement="left"
-                                  onClick={() =>setShowMPAInfo(true)}
+                                  onClick={() => setShowMPAInfo(true)}
                                 />
                                 <Modal
                                   isOpen={showMPAInfo}
-                                  onClose={() =>setShowMPAInfo(false)}
+                                  onClose={() => setShowMPAInfo(false)}
                                   title={layerTitle(layer)}
-                                  >
-                                    {t('map.descriptionMPA', 'Marine protected areas (MPAs)...')}
-      
+                                >
+                                  {t('map.descriptionMPA', 'Marine protected areas (MPAs)...')}
                                 </Modal>
                               </div>
                             )}