+++ packages/api-types/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @globalfishingwatch/api-types
 
+## 6.2.0
+
+### Minor Changes
+
+- 2ea27a38: support draw geometryType in dataset configuration
+
+## 6.1.0
+
+### Minor Changes
+
+- ec34c600: added amount of transmissions
+
 ## 6.0.1
 
 ### Patch Changes